+++ asana_extensions/rules/__init__.py
@@ -1,6 +1,6 @@
 # pylint: disable=missing-module-docstring
 __all__ = [
-        'config',
-        'dirs',
-        'exceptions',
+    'move_tasks_rule',
+    'rule_meta',
+    'rules',
 ]