+++ tests/unit/rules/__init__.py
@@ -1,0 +1,6 @@
+# pylint: disable=missing-module-docstring
+__all__ = [
+        'test_move_tasks_rule',
+        'test_rule_meta',
+        'test_rules',
+]